+++ frigate/test/test_http.py
@@ -122,7 +122,6 @@ class TestHttp(unittest.TestCase):
             None,
             None,
             None,
-            None,
         )
         id = "123456.random"
 
@@ -144,7 +143,6 @@ class TestHttp(unittest.TestCase):
             None,
             None,
             None,
-            None,
         )
         id = "123456.random"
         bad_id = "654321.other"
@@ -159,7 +157,6 @@ class TestHttp(unittest.TestCase):
         app = create_fastapi_app(
             FrigateConfig(**self.minimal_config),
             self.db,
-            None,
             None,
             None,
             None,
@@ -188,7 +185,6 @@ class TestHttp(unittest.TestCase):
             None,
             None,
             None,
-            None,
         )
         id = "123456.random"
 
@@ -209,7 +205,6 @@ class TestHttp(unittest.TestCase):
         app = create_fastapi_app(
             FrigateConfig(**self.minimal_config),
             self.db,
-            None,
             None,
             None,
             None,
@@ -249,7 +244,6 @@ class TestHttp(unittest.TestCase):
         app = create_fastapi_app(
             FrigateConfig(**self.minimal_config),
             self.db,
-            None,
             None,
             None,
             None,
@@ -300,7 +294,6 @@ class TestHttp(unittest.TestCase):
             None,
             None,
             None,
-            None,
             mock_event_updater,
         )
         id = "123456.random"
@@ -334,7 +327,6 @@ class TestHttp(unittest.TestCase):
             None,
             None,
             None,
-            None,
         )
 
         with TestClient(app) as client:
@@ -346,7 +338,6 @@ class TestHttp(unittest.TestCase):
         app = create_fastapi_app(
             FrigateConfig(**self.minimal_config),
             self.db,
-            None,
             None,
             None,
             None,
