+++ process_clip.py
@@ -15,8 +15,8 @@ sys.path.append("/workspace/frigate")
 from frigate.config import FrigateConfig  # noqa: E402
 from frigate.motion import MotionDetector  # noqa: E402
 from frigate.object_detection import LocalObjectDetector  # noqa: E402
-from frigate.object_processing import CameraState  # noqa: E402
 from frigate.track.centroid_tracker import CentroidTracker  # noqa: E402
+from frigate.track.object_processing import CameraState  # noqa: E402
 from frigate.util import (  # noqa: E402
     EventsPerSecond,
     SharedMemoryFrameManager,
